# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OutreachHub API
  description: |
    OutreachHub 智能拓客与邮件营销平台 API。

    ## 认证
    所有 API 请求需要通过 `Authorization: Bearer <token>` header 或 `auth-token` cookie 进行认证。

    ## 通用响应格式
    - 成功：`{ "success": true, "data": ... }`
    - 错误：`{ "success": false, "error": { "code": "...", "message": "..." } }`
    - 分页：`{ "success": true, "data": [...], "pagination": { "page": 1, "limit": 20, "total": 100, "pages": 5 } }`
  version: 1.0.0
  contact:
    name: OutreachHub Support
    email: support@outreachhub.com
servers:
  - url: http://localhost:3030
    description: 本地开发
  - url: https://app.outreachhub.com
    description: 生产环境
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: 'API Key (格式: oh_xxx)'
  responses:
    RateLimitUnavailable:
      description: API Key 分布式限流服务暂不可用
      headers:
        Retry-After:
          description: 建议重试等待秒数
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    Contact:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        title:
          type: string
        companyId:
          type: string
        status:
          type: string
          enum: [NEW, CONTACTED, INTERESTED, NOT_INTERESTED, QUALIFIED, CONVERTED, UNREACHABLE]
        tags:
          type: array
          items:
            type: string
        emails:
          type: array
          items:
            $ref: '#/components/schemas/ContactEmail'
        createdAt:
          type: string
          format: date-time
    ContactEmail:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        type:
          type: string
          enum: [PROFESSIONAL, PERSONAL]
        isVerified:
          type: boolean
...
paths:
...
  /api/contacts/{id}:
    get:
      summary: 获取联系人详情
      tags: [联系人]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
        '404':
          description: 联系人不存在
    put:
      summary: 更新联系人
      tags: [联系人]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除联系人
      tags: [联系人]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
...
  /api/v1/contacts/{id}:
    get:
      summary: 获取联系人详情（公开 API v1）
      description: |
        需要 `contacts:read` 或 `contacts:manage` 权限。
        自动过滤为当前租户数据，无法访问其他租户联系人。
      tags: [公开 API v1]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Contact'
        '401':
          description: 未认证或 API Key 无效
        '403':
          description: 权限不足
        '404':
          description: 联系人不存在（或不属于当前租户）
        '429':
          description: 请求频率超出限制
        '503':
          $ref: '#/components/responses/RateLimitUnavailable'
    put:
      summary: 更新联系人（公开 API v1）
      description: |
        需要 `contacts:write` 或 `contacts:manage` 权限。
        自动过滤为当前租户数据。
      tags: [公开 API v1]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                title:
                  type: string
                companyId:
                  type: string
                emails:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
                notes:
                  type: string
                status:
                  type: string
                  enum: [NEW, CONTACTED, INTERESTED, NOT_INTERESTED, QUALIFIED, CONVERTED, UNREACHABLE]
      responses:
        '200':
          description: 更新成功
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Contact'
        '401':
          description: 未认证或 API Key 无效
        '403':
          description: 权限不足
        '404':
          description: 联系人不存在
        '429':
          description: 请求频率超出限制
        '503':
          $ref: '#/components/responses/RateLimitUnavailable'
    delete:
      summary: 删除联系人（公开 API v1）
      description: |
        需要 `contacts:write` 或 `contacts:manage` 权限。
        级联删除关联的邮箱、邮件日志和活动联系人记录。
      tags: [公开 API v1]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
        '401':
          description: 未认证或 API Key 无效
        '403':
          description: 权限不足
        '404':
          description: 联系人不存在
        '429':
          description: 请求频率超出限制
        '503':
          $ref: '#/components/responses/RateLimitUnavailable'
  /api/api-keys:
    get:
      summary: 获取 API Key 列表
      tags: [API Key 管理]
      description: 返回租户所有 API Key（仅显示前缀，不显示完整密钥）
      responses:
        '200':
          description: 成功
    post:
      summary: 创建 API Key
      tags: [API Key 管理]
      description: 创建新的 API Key，完整密钥仅在响应中返回一次
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
                    enum: [contacts:read, contacts:write, campaigns:read, campaigns:write]
                rateLimit:
                  type: integer
                  default: 100
      responses:
        '201':
          description: 创建成功
  /api/api-keys/{id}:
    patch:
      summary: 更新 API Key
      tags: [API Key 管理]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 吊销 API Key
      tags: [API Key 管理]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 吊销成功
...